# docs/contracts/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AdminReviewResponse": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "customer_email": {
            "title": "Customer Email",
            "type": "string"
          },
          "customer_name": {
            "title": "Customer Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "internal_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reason"
          },
          "menu_item_name": {
            "title": "Menu Item Name",
            "type": "string"
          },
          "menu_item_slug": {
            "title": "Menu Item Slug",
            "type": "string"
          },
          "moderated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderated At"
          },
          "moderated_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderated By Name"
          },
          "order_item_id": {
            "format": "uuid",
            "title": "Order Item Id",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_item_id",
          "menu_item_slug",
          "menu_item_name",
          "rating",
          "body",
          "status",
          "created_at",
          "updated_at",
          "customer_name",
          "customer_email",
          "internal_reason",
          "moderated_by_name",
          "moderated_at"
        ],
        "title": "AdminReviewResponse",
        "type": "object"
      },
...
      "AllergenResponse": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "note"
        ],
        "title": "AllergenResponse",
        "type": "object"
      },
...
      "AvailabilityState": {
        "enum": [
          "available",
          "temporarily_unavailable",
          "scheduled"
        ],
        "title": "AvailabilityState",
        "type": "string"
      },
...
      "CustomerReviewEligibilityResponse": {
        "properties": {
          "menu_item_slugs": {
            "items": {
              "type": "string"
            },
            "title": "Menu Item Slugs",
            "type": "array"
          }
        },
        "required": [
          "menu_item_slugs"
        ],
        "title": "CustomerReviewEligibilityResponse",
        "type": "object"
      },
      "CustomerReviewResponse": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "menu_item_name": {
            "title": "Menu Item Name",
            "type": "string"
          },
          "menu_item_slug": {
            "title": "Menu Item Slug",
            "type": "string"
          },
          "order_item_id": {
            "format": "uuid",
            "title": "Order Item Id",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_item_id",
          "menu_item_slug",
          "menu_item_name",
          "rating",
          "body",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "CustomerReviewResponse",
        "type": "object"
      },
      "CustomerReviewUpdateRequest": {
        "properties": {
          "body": {
            "maxLength": 1500,
            "minLength": 12,
            "title": "Body",
            "type": "string"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "rating",
          "body"
        ],
        "title": "CustomerReviewUpdateRequest",
        "type": "object"
      },
      "CustomerReviewWriteRequest": {
        "properties": {
          "body": {
            "maxLength": 1500,
            "minLength": 12,
            "title": "Body",
            "type": "string"
          },
          "order_item_id": {
            "format": "uuid",
            "title": "Order Item Id",
            "type": "string"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "order_item_id",
          "rating",
          "body"
        ],
        "title": "CustomerReviewWriteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MenuItemResponse": {
        "properties": {
          "allergens": {
            "items": {
              "$ref": "#/components/schemas/AllergenResponse"
            },
            "title": "Allergens",
            "type": "array"
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityState"
          },
          "base_price_minor": {
            "title": "Base Price Minor",
            "type": "integer"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryResponse"
          },
          "currency_code": {
            "title": "Currency Code",
            "type": "string"
          },
          "demo_discount_minor": {
            "title": "Demo Discount Minor",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dietary_tags": {
            "items": {
              "type": "string"
            },
            "title": "Dietary Tags",
            "type": "array"
          },
          "display_order": {
            "title": "Display Order",
            "type": "integer"
          },
          "final_price_minor": {
            "title": "Final Price Minor",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "media": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "option_groups": {
            "items": {
              "$ref": "#/components/schemas/OptionGroupResponse"
            },
            "title": "Option Groups",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "ingredients",
          "dietary_tags",
          "base_price_minor",
          "demo_discount_minor",
          "final_price_minor",
          "currency_code",
          "display_order",
          "category",
          "availability",
          "media",
          "allergens",
          "option_groups"
        ],
        "title": "MenuItemResponse",
        "type": "object"
      },
...
      "PublicReviewListResponse": {
        "properties": {
          "average_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Rating"
          },
          "review_count": {
            "title": "Review Count",
            "type": "integer"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/PublicReviewResponse"
            },
            "title": "Reviews",
            "type": "array"
          }
        },
        "required": [
          "review_count",
          "average_rating",
          "reviews"
        ],
        "title": "PublicReviewListResponse",
        "type": "object"
      },
...
      "ReviewModerationAction": {
        "enum": [
          "approve",
          "reject",
          "hide",
          "restore"
        ],
        "title": "ReviewModerationAction",
        "type": "string"
      },
      "ReviewModerationRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ReviewModerationAction"
          },
          "internal_reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reason"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewModerationRequest",
        "type": "object"
      },
      "ReviewStatus": {
        "enum": [
          "pending",
          "approved",
          "rejected",
          "hidden"
        ],
        "title": "ReviewStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Local direct-ordering demo API.",
    "title": "Nosh API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/account/reviews": {
      "get": {
        "operationId": "list_customer_reviews_api_v1_account_reviews_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CustomerReviewResponse"
                  },
                  "title": "Response List Customer Reviews Api V1 Account Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Customer Reviews",
        "tags": [
          "Customer reviews"
        ]
      },
      "post": {
        "operationId": "create_customer_review_api_v1_account_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerReviewWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Customer Review",
        "tags": [
          "Customer reviews"
        ]
      }
    },
    "/api/v1/account/reviews/eligible-menu-items": {
      "get": {
        "operationId": "list_eligible_review_menu_items_api_v1_account_reviews_eligible_menu_items_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerReviewEligibilityResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Eligible Review Menu Items",
        "tags": [
          "Customer reviews"
        ]
      }
    },
    "/api/v1/account/reviews/{review_id}": {
      "delete": {
        "operationId": "delete_customer_review_api_v1_account_reviews__review_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Customer Review",
        "tags": [
          "Customer reviews"
        ]
      },
      "patch": {
        "operationId": "update_customer_review_api_v1_account_reviews__review_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerReviewUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Customer Review",
        "tags": [
          "Customer reviews"
        ]
      }
    },
...
    "/api/v1/admin/reviews": {
      "get": {
        "operationId": "list_admin_reviews_api_v1_admin_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminReviewResponse"
                  },
                  "title": "Response List Admin Reviews Api V1 Admin Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Admin Reviews",
        "tags": [
          "Admin review moderation"
        ]
      }
    },
    "/api/v1/admin/reviews/{review_id}/moderation": {
      "post": {
        "operationId": "moderate_review_api_v1_admin_reviews__review_id__moderation_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewModerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Moderate Review",
        "tags": [
          "Admin review moderation"
        ]
      }
    },
...
    "/api/v1/catalog/menu-items": {
      "get": {
        "operationId": "list_menu_items_api_v1_catalog_menu_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "category_slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z0-9-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Slug"
            }
          },
          {
            "in": "query",
            "name": "location_slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z0-9-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MenuItemResponse"
                  },
                  "title": "Response List Menu Items Api V1 Catalog Menu Items Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List published menu items",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/menu-items/{slug}": {
      "get": {
        "operationId": "read_menu_item_api_v1_catalog_menu_items__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "location_slug",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z0-9-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a published menu item",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/menu-items/{slug}/reviews": {
      "get": {
        "operationId": "list_public_reviews_api_v1_catalog_menu_items__slug__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9-]+$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Reviews",
        "tags": [
          "Catalog"
        ]
      }
    },
...
  }
}